+++ test/baselines/connectionIndicator_yes/src/main.ts
@@ -1,37 +1,40 @@
-"use strict";
-module.exports = (async (answers) => {
-    const useJavaScript = answers.language === "JavaScript";
-    const useES6Class = answers.es6class === "yes";
-    if (!useJavaScript || useES6Class)
-        return;
-    const template = `
-"use strict";
-
 /*
- * Created with @iobroker/create-adapter v${answers.creatorVersion}
+ * Created with @iobroker/create-adapter v1.8.0
  */
 
 // The adapter-core module gives you access to the core ioBroker functions
 // you need to create an adapter
-const utils = require("@iobroker/adapter-core");
+import * as utils from "@iobroker/adapter-core";
 
 // Load your modules here, e.g.:
-// const fs = require("fs");
+// import * as fs from "fs";
 
-/**
- * The adapter instance
- * @type {ioBroker.Adapter}
- */
-let adapter;
+// Augment the adapter.config object with the actual types
+// TODO: delete this in the next version
+declare global {
+	namespace ioBroker {
+		interface AdapterConfig {
+			// Define the shape of your options here (recommended)
+			option1: boolean;
+			option2: string;
+			// Or use a catch-all approach
+			[key: string]: any;
+		}
+	}
+}
+
+let adapter: ioBroker.Adapter;
 
 /**
  * Starts the adapter instance
- * @param {Partial<ioBroker.AdapterOptions>} [options]
  */
-function startAdapter(options) {
+function startAdapter(options: Partial<ioBroker.AdapterOptions> = {}) {
 	// Create the adapter and define its methods
-	return adapter = utils.adapter(Object.assign({}, options, {
-		name: "${answers.adapterName}",
+	return adapter = utils.adapter({
+		// Default options
+		...options,
+		// custom options
+		name: "test-adapter",
 
 		// The ready callback is called when databases are connected and adapter received configuration.
 		// start here!
@@ -51,10 +54,10 @@ function startAdapter(options) {
 		objectChange: (id, obj) => {
 			if (obj) {
 				// The object was changed
-				adapter.log.info(\`object \$\{id} changed: \$\{JSON.stringify(obj)}\`);
+				adapter.log.info(`object ${id} changed: ${JSON.stringify(obj)}`);
 			} else {
 				// The object was deleted
-				adapter.log.info(\`object \$\{id} deleted\`);
+				adapter.log.info(`object ${id} deleted`);
 			}
 		},
 
@@ -62,10 +65,10 @@ function startAdapter(options) {
 		stateChange: (id, state) => {
 			if (state) {
 				// The state was changed
-				adapter.log.info(\`state \$\{id} changed: \$\{state.val} (ack = \$\{state.ack})\`);
+				adapter.log.info(`state ${id} changed: ${state.val} (ack = ${state.ack})`);
 			} else {
 				// The state was deleted
-				adapter.log.info(\`state \$\{id} deleted\`);
+				adapter.log.info(`state ${id} deleted`);
 			}
 		},
 
@@ -82,15 +85,13 @@ function startAdapter(options) {
 		// 		}
 		// 	}
 		// },
-	}));
+	});
 }
 
 function main() {
 
-${answers.connectionIndicator === "yes" ? `
 	// Reset the connection indicator during startup
 	this.setState("info.connection", false, true);
-` : ""}
 
 	// The adapters config (in the instance object everything under the attribute "native") is accessible via
 	// adapter.config:
@@ -148,6 +149,3 @@ if (module.parent) {
 	// otherwise start the instance directly
 	startAdapter();
 }
-`;
-    return template.trim();
-});