+++ travis/create_templates.ts
@@ -30,6 +30,7 @@ const adapterAnswers: Answers = {
 	...baseAnswers,
 	startMode: "daemon",
 	features: ["adapter"],
+	connectionIndicator: "no",
 	es6class: "yes",
 	type: "general",
 	adminFeatures: ["custom", "tab"],